Updated dev

+++ lib/Model/responseClass.dart
@@ -130,8 +130,8 @@ class Main {
   double feelsLike;
   double tempMin;
   double tempMax;
-  int pressure;
-  int humidity;
+  double pressure;
+  double humidity;
 
   Main(
       {this.temp,
@@ -142,12 +142,12 @@ class Main {
       this.humidity});
 
   Main.fromJson(Map<String, dynamic> json) {
-    temp = json['temp'];
-    feelsLike = json['feels_like'];
-    tempMin = json['temp_min'];
-    tempMax = json['temp_max'];
-    pressure = json['pressure'];
-    humidity = json['humidity'];
+    temp = json['temp'].toDouble();
+    feelsLike = json['feels_like'].toDouble();
+    tempMin = json['temp_min'].toDouble();
+    tempMax = json['temp_max'].toDouble();
+    pressure = json['pressure'].toDouble();
+    humidity = json['humidity'].toDouble();
   }
 
   Map<String, dynamic> toJson() {
